+++ packages/postcss-tape/src/noop-plugin.ts
@@ -1,0 +1,12 @@
+const noopPlugin = () => {
+	return {
+		postcssPlugin: 'noop-plugin',
+		Rule() {
+			// do nothing
+		},
+	};
+};
+
+noopPlugin.postcss = true;
+
+export default noopPlugin;