+++ plugins/postcss-custom-properties/test/export-properties.mjs
@@ -15,5 +15,6 @@ export const customProperties = {
 	'--url-2': 'url(\'/my/path\')',
 	'--url-3': 'url(/my/path)',
 	'--url-4': 'url(data:image/png;bm90LWFuZC1pbWFnZQ==)',
+	'--✅-size': '2em',
 	'--theme-color': '#053'
 };