+++ plugins/postcss-custom-properties/test/import-properties-2.mjs
@@ -1,0 +1,6 @@
+export default {
+	customProperties: {
+		'--color': 'rgb(255, 0, 0)',
+		'--color-2': 'yellow',
+	},
+};