+++ plugin-packs/postcss-preset-env/CHANGELOG.md
@@ -2,10 +2,8 @@
 
 ### Unreleased (minor)
 
-- Added `@csstools/postcss-is-pseudo-class`
+- Added `@csstools/postcss-is-pseudo-class` <br/> [Check the plugin README](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-is-pseudo-class#readme) for usage details.
 - Added `debug` [option](https://github.com/csstools/postcss-plugins/tree/main/plugin-packs/postcss-preset-env#debug) that enables extra debugging information while processing the CSS.
-
-[check the plugin readme](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-is-pseudo-class#readme) for usage details.
 
 ### 7.2.3 (January 12, 2022)